requirements.txt
aiogram==3.15.0
groq==0.13.0
google-genai==0.3.0
gTTS==2.5.4
python-dotenv==1.0.1
pydub==0.25.1
